added chilren prop in chatroom component
